Modified to write diffimage_uncertainty FITS image as 4-byte float data.

+++ pipeline/awsBatchSubmitJobs_runSingleSciencePipeline.py
@@ -231,7 +231,7 @@ def compute_diffimage_uncertainty(sca_gain,
 
     hdu_list_unc = []
     data_unc = np.sqrt((pos_np_data_sci + pos_np_data_ref) / sca_gain + std_dif_img * std_dif_img)
-    hdu_unc = fits.PrimaryHDU(header=hdr_sci,data=data_unc)
+    hdu_unc = fits.PrimaryHDU(header=hdr_sci,data=data_unc.astype(np.float32))
     hdu_list_unc.append(hdu_unc)
     hdu_unc = fits.HDUList(hdu_list_unc)
     hdu_unc.writeto(diffimage_unc_filename,overwrite=True,checksum=True)
